+++ testbed/scenes/concavemesh/ConcaveMeshScene.cpp
@@ -108,9 +108,6 @@ ConcaveMeshScene::ConcaveMeshScene(const std::string& name, EngineSettings& sett
         // Create a sphere and a corresponding rigid in the physics world
         Sphere* sphere = new Sphere(true, SPHERE_RADIUS, mPhysicsCommon, mPhysicsWorld, meshFolderPath);
 
-        // Add some rolling resistance
-        sphere->getCollider()->getMaterial().setRollingResistance(rp3d::decimal(0.08));
-
         // Set the box color
         sphere->setColor(mObjectColorDemo);
         sphere->setSleepingColor(mSleepingColorDemo);
@@ -130,8 +127,6 @@ ConcaveMeshScene::ConcaveMeshScene(const std::string& name, EngineSettings& sett
         // Create a cylinder and a corresponding rigid in the physics world
         Capsule* capsule = new Capsule(true, CAPSULE_RADIUS, CAPSULE_HEIGHT,
                                        mPhysicsCommon, mPhysicsWorld, meshFolderPath);
-
-        capsule->getCollider()->getMaterial().setRollingResistance(rp3d::decimal(0.08));
 
         // Set the box color
         capsule->setColor(mObjectColorDemo);